requirements.txt
Flask
Flask-Cors
opencv-python-headless
numpy
facenet-pytorch
tqdm
Pillow
werkzeug
scikit-learn